Minor update to footer and embed fix
# Portfolio Site

This is a website that will function as my portfolio for my study.

## Updates
- Added a missing space in the footer
- Fixed the pdf embed (capitalisation issue broke it in the most recent version)

This version is not yet hosted online. With the exception of the problematic embed issue, these changes will be coming to the next version.
The pdf embed was manually fixed in the web version.

## Roadmap
- Add project and portfolio content properly, incl text and images
- I think I want to decorate the footer a little more.
- If I can find the time to, I want to add more little details to enhance the early 2000s personal website/myspace page feel.
- Images on the site are currently not optimised. I might want to optimise them for web use and implement them as webp files instead.
- The sounds for the Home and Portfolio buttons in the navbar stopped working when I added a link to them. That's a shame. I'll try see if i can fix this.
- There are a couple of warnings i need to get rid of

## Contributing
I currently am not open to contributions. This is a personal project.